Add wasEqualsRecentlyPressed and equalsButtonPressed
Adds function wasEqualsRecentlyPressed to check whether the equals
button was recently pressed and if the screen needs to be cleared. Also
adds equalsButtonPressed function to replace an existing function that
served as the event handler for the equals button as the previous one
was not compatible with the new variable that tracks the equal button.